support `arrays_zip` function
